==> 02_vector_db.ipynb <==
import pandas as pd
from langchain.embeddings.sentence_transformer import SentenceTransformerEmbeddings
from langchain.vectorstores import Chroma
from langchain.schema import Document
import os





    
df = pd.read_csv('./dataset/va_content.csv')

documents = [Document(page_content=row['title'] +'/n' + row['content'], metadata={'source_url': row['source_url']}) for index, row in df.iterrows()]

from langchain.embeddings.sentence_transformer import SentenceTransformerEmbeddings
embedding_function = SentenceTransformerEmbeddings(model_name="all-MiniLM-L6-v2")
vectordb = Chroma.from_documents(documents, embedding_function,persist_directory="./chroma_db_alt")
vectordb.persist()

print("Vector store created and saved locally.")

# Load the vector store for similarity search
vectordb = Chroma(persist_directory="./chroma_db_alt", embedding_function=embedding_function)

def similarity_search(query, k=3):
    results = vectordb.similarity_search(query, k=k)
    return results



# Test the similarity search
query = "What shuold I do after VA disability claim?"
results = similarity_search(query)

results
