==> 03_RAG_LLM.ipynb <==
# Try different LLM for RAG

import pandas as pd
from langchain.embeddings.sentence_transformer import SentenceTransformerEmbeddings
from langchain.vectorstores import Chroma
from langchain.schema import Document
from transformers import AutoTokenizer, AutoModelForCausalLM, pipeline
import time
import torch


!nvidia-smi

# !sudo kill -9 38817

# Load chromadb and csv
from langchain.embeddings.sentence_transformer import SentenceTransformerEmbeddings
embedding_function = SentenceTransformerEmbeddings(model_name="all-MiniLM-L6-v2")

vectordb = Chroma(persist_directory="./chroma_db_alt", embedding_function=embedding_function)

# Function to perform similarity search
def similarity_search(query, k=5):
    results = vectordb.similarity_search(query, k=k)
    return results




# Load LLaMA 3 model
model_id = "meta-llama/Meta-Llama-Guard-2-8B"
device = "cuda" if torch.cuda.is_available() else "cpu"


# Initialize model
tokenizer = AutoTokenizer.from_pretrained(model_id)
model = AutoModelForCausalLM.from_pretrained(model_id, torch_dtype=torch.float32).to(device)


# Basic llama3

# Function to generate answers using the basic LLM
def generate_basic_llm_answer(question):
    prompt = (
        f"You are an expert on VA disability benefits. You have deep knowledge and understanding of all aspects related to VA disability claims and benefits. "
        f"Please provide a detailed and accurate answer to the following question based on your expertise.\n\n"
        f"Question: {question}\n"
        f"Answer:"
    )
    
    input_ids = tokenizer(prompt, return_tensors="pt").input_ids.to(device)
    output = model.generate(input_ids=input_ids, max_new_tokens=500, pad_token_id=tokenizer.pad_token_id, temperature=0)
    
    generated_text = tokenizer.decode(output[0], skip_special_tokens=True)
    
    answer = generated_text.split("Answer:")[1].strip() if "Answer:" in generated_text else generated_text.strip()
    
    return answer

def log_process_time(start_time):
    end_time = time.time()
    return end_time - start_time

def process_csv(df):
    
    answers = []
    times = []
    
    for question in df['question']:
        start_time = time.time()
        answer = generate_basic_llm_answer(question)
        process_time = log_process_time(start_time)
        
        answers.append(answer)
        times.append(process_time)
    
    df['basic_llm'] = answers
    df['basic_time'] = times
    
    return df

# LLama3 RAG


# LLama3 RAG


# generate answers using RAG
def generate_rag_answer(question, k=3):
    results = similarity_search(question, k)
    
    context = "\n\n".join([f"Source {i+1}: {result.page_content}" for i, result in enumerate(results)])
    
    prompt = (
        f"You are an expert on VA disability benefits. You will be provided with context information from several sources.\n"
        f"Use only the provided information to answer the question accurately and concisely. Do not use any external knowledge. "
        f"Base your answer strictly on the context provided.\n\n"
        f"Please follow these instructions carefully to provide an accurate answer:\n"
        f"1. Carefully read each paragraph of the provided content.\n"
        f"2. Identify if the paragraph contains relevant information to answer the question. If a paragraph does not provide relevant information, move to the next paragraph.\n"
        f"3. When you find relevant information, use it to construct your answer. Include as much evidence as possible from the context to support your answer, even if an answer has already been started.\n"
        f"4. Ensure your answer is accurate, concise, and based solely on the provided context.\n\n"
        f"Context:\n"
        f"{context}\n\n"
        f"Question: {question}\n"
        f"Answer:"
    )
    
    input_ids = tokenizer(prompt, return_tensors="pt").input_ids.to(device)
    output = model.generate(input_ids=input_ids, max_new_tokens=500, pad_token_id=tokenizer.pad_token_id, temperature=0)
    
    generated_text = tokenizer.decode(output[0], skip_special_tokens=True)
    
    answer = generated_text.split("Answer:")[1].strip() if "Answer:" in generated_text else generated_text.strip()
    
    return answer, results

def log_process_time(start_time):
    end_time = time.time()
    return end_time - start_time

def process_rag_csv(df):
    
    rag_answers = []
    process_times = []
    similarity_metadata = []

    for question in df['question']:
        start_time = time.time()
        answer, results = generate_rag_answer(question)
        process_time = log_process_time(start_time)
        
        similarity_info = "\n\n".join([f"Title: {result.metadata['source_url']}\nContent: {result.page_content}" for result in results])
        
        rag_answers.append(answer)
        process_times.append(process_time)
        similarity_metadata.append(similarity_info)
    
    df['rag_llm'] = rag_answers
    df['rag_time'] = process_times
    df['rag_similarity'] = similarity_metadata
    
    return df




df_va_content = pd.read_csv('./dataset/processed_va_questions.csv')
df_va_content =  df_va_content.dropna(subset=['answer'])
df_va_content.head()

df_process = process_csv(df_va_content)

df_rag = process_rag_csv(df_process)

df_rag.to_csv('./dataset/qa_va_content_rag.csv')

df_rag.head()

results[0]

# Rag With Reranker

import pandas as pd
import torch
from transformers import AutoTokenizer, AutoModelForCausalLM
from sentence_transformers import SentenceTransformer, util
import numpy as np
from llama_index.core import VectorStoreIndex,SimpleDirectoryReader,ServiceContext,PromptTemplate
from sentence_transformers import SentenceTransformer, util
from langchain.vectorstores import Chroma
from llama_index.core.postprocessor import SentenceTransformerRerank

model_id = "meta-llama/Meta-Llama-Guard-2-8B"
device = "cuda" if torch.cuda.is_available() else "cpu"

tokenizer = AutoTokenizer.from_pretrained(model_id)
model = AutoModelForCausalLM.from_pretrained(model_id, torch_dtype=torch.float32).to(device)

embedding_model = SentenceTransformer('all-MiniLM-L6-v2')

vectordb = Chroma(persist_directory="./chroma_db_alt", embedding_function=embedding_model.encode)

reranker = SentenceTransformerRerank()

def similarity_search(query, k=5):
    results = vectordb.similarity_search(query, k=k)
    return results

def generate_rag_answer(question, k=5):
    results = similarity_search(question, k)
    
    reranked_results = reranker.rerank(results, query=question)
    
    context = "\n\n".join([f"Source {i+1}: {result['page_content']}" for i, result in enumerate(reranked_results)])
    
    prompt = (
        f"You are an expert on VA disability benefits. You will be provided with context information from several sources.\n"
        f"Use only the provided information to answer the question accurately and concisely. Do not use any external knowledge. "
        f"Base your answer strictly on the context provided.\n\n"
        f"Please follow these instructions carefully to provide an accurate answer:\n"
        f"1. Carefully read each paragraph of the provided content.\n"
        f"2. Identify if the paragraph contains relevant information to answer the question. If a paragraph does not provide relevant information, move to the next paragraph.\n"
        f"3. When you find relevant information, use it to construct your answer. Include as much evidence as possible from the context to support your answer, even if an answer has already been started.\n"
        f"4. Ensure your answer is accurate, concise, and based solely on the provided context.\n\n"
        f"Context:\n"
        f"{context}\n\n"
        f"Question: {question}\n"
        f"Answer:"
    )
    
    input_ids = tokenizer(prompt, return_tensors="pt").input_ids.to(device)
    output = model.generate(input_ids=input_ids, max_new_tokens=500, pad_token_id=tokenizer.pad_token_id)
    
    generated_text = tokenizer.decode(output[0], skip_special_tokens=True)
    
    answer = generated_text.split("Answer:")[1].strip() if "Answer:" in generated_text else generated_text.strip()
    
    return answer, reranked_results


def process_rag_rerank_csv(df):
    
    rag_answers = []
    process_times = []
    similarity_metadata = []

    for question in df['question']:
        start_time = time.time()
        answer, results = generate_rag_answer(question)
        process_time = log_process_time(start_time)
        
        similarity_info = "\n\n".join([f"Title: {result.metadata['source_url']}\nContent: {result.page_content}" for result in results])
        
        rag_answers.append(answer)
        process_times.append(process_time)
        similarity_metadata.append(similarity_info)
    
    df['rag_rerank_llm'] = rag_answers
    df['rag_rerank_time'] = process_times
    df['rag_rerank_similarity'] = similarity_metadata
    
    return df

df_rerank = process_rag_rerank_csv(df_rag)


df_rerank.to_csv('./dataset/qa_va_content_rag.csv')
